LayoutTests/accessibility/mac/attributed-string/attributed-string-for-range-with-options.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/accessibility-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body>

<div id="content" contenteditable="true">word mizpelled word</div>

<script>
if (window.accessibilityController) {
    let output = "This test ensures that attributed string for text marker range works and only includes misspelled attribute when requested.\n\n";

    var text = null;
    var expected1 = "AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, word mizpelled word";
    var expected2 = "AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, Misspelled, AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, word mizpelled word";
    let axElement = accessibilityController.accessibleElementById("content");
    let range = axElement.textMarkerRangeForElement(axElement);
    text = axElement.attributedStringForTextMarkerRangeWithOptions(range, false);
    output += expect("text", "expected1");
    text = axElement.attributedStringForTextMarkerRangeWithOptions(range, true);
    output += expect("text", "expected2");

    // Correct the misspelling and verify that the attributed string changes accordingly.
    var expected3 = "AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, word misspelled word";
    axElement.replaceTextInRange("ss", 7, 1);
    // Note that with replaceTextInRange(...), there is no need to await for the change to be reflected in the AX tree.
    range = axElement.textMarkerRangeForElement(axElement);
    text = axElement.attributedStringForTextMarkerRangeWithOptions(range, true);
    output += expect("text", "expected3");

    // Make another misspelling and verify.
    var expected4 = "Misspelled, AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, AXFont - {\n    AXFontFamily = Times;\n    AXFontName = \"Times-Roman\";\n    AXFontSize = 16;\n}, worz misspelled word";
    axElement.replaceTextInRange("z", 3, 1);
    range = axElement.textMarkerRangeForElement(axElement);
    text = axElement.attributedStringForTextMarkerRangeWithOptions(range, true);
    output += expect("text", "expected4");

    document.getElementById("content").style.visibility = "hidden";
    debug(output);
}
</script>
</body>
</html>
